+++ sources/System/Renderer/CameraSystem.cpp
@@ -53,8 +53,8 @@ namespace BBM
 			introAnimation(entity, true);
 		if (playerPos.size() == 1)
 			newCameraPos = playerPos[0];
-		for (int i = 0; i < static_cast<int>(playerPos.size()); i++)
-			for (int j = 0; j < static_cast<int>(playerPos.size()); j++) {
+		for (size_t i = 0; i < playerPos.size(); i++)
+			for (size_t j = 0; j < playerPos.size(); j++) {
 				if (maxDist < playerPos[i].distance(playerPos[j])) {
 					maxDist = playerPos[i].distance(playerPos[j]);
 					newCameraPos = (playerPos[i] + playerPos[j]) / 2;
